requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
werkzeug==3.0.1
requests==2.32.3
python-dotenv==1.0.0
psycopg2-binary
